Resolve compiler warnings due to sampler code
Compilation of the sampler resulted in quite a few compiler warnings.
This commit resolves the warnings caused by the sampler code.
